src/handler/email/remove-email.ts
import { Context } from "../../typing";
import { ClientError } from "@lindorm-io/errors";

interface Options {
  identityId: string;
  email: string;
}

export const removeEmail = async (ctx: Context, options: Options): Promise<void> => {
  const {
    repository: { emailRepository },
  } = ctx;

  const { identityId, email } = options;

  const entity = await emailRepository.find({ identityId, email });

  if (entity.primary) {
    throw new ClientError("Unable to remove primary email");
  }

  await emailRepository.remove(entity);
};
