src/handler/email/remove-email.test.ts
import { ClientError } from "@lindorm-io/errors";
import { Email } from "../../entity";
import { removeEmail } from "./remove-email";

describe("removeEmail", () => {
  const email1 = new Email({
    identityId: "identityId",
    email: "email",
    primary: false,
    verified: true,
  });

  let ctx: any;
  let options: any;

  beforeEach(() => {
    ctx = {
      repository: {
        emailRepository: {
          find: jest.fn().mockResolvedValue(email1),
          remove: jest.fn(),
        },
      },
    };
    options = {
      identityId: "identityId",
      email: "email2",
    };
  });

  test("should remove email", async () => {
    await expect(removeEmail(ctx, options)).resolves.toBeUndefined();

    expect(ctx.repository.emailRepository.remove).toHaveBeenCalledWith(email1);
  });

  test("should throw when email is primary", async () => {
    email1.primary = true;

    await expect(removeEmail(ctx, options)).rejects.toThrow(ClientError);
  });
});
